requirement.txt
Django==1.11.7
django-pure-pagination==0.3.0
django-simple-captcha==0.4.6
DjangoUeditor==1.8.143
MySQL-python==1.2.5
olefile==0.44
Pillow==4.3.0
pip==9.0.1
pytz==2017.3
setuptools==38.1.0
six==1.11.0
wheel==0.30.0
